+++ omnigibson/objects/controllable_object.py
@@ -383,7 +383,7 @@ class ControllableObject(BaseObject):
         u_type_vec = th.tensor([ControlType.NONE] * self.n_dof)
         for group, ctrl in control.items():
             idx = self._controllers[group].dof_idx
-            u_vec[idx] = ctrl["value"]
+            u_vec[idx] = ctrl["value"].float()
             u_type_vec[idx] = ctrl["type"]
 
         u_vec, u_type_vec = self._postprocess_control(control=u_vec, control_type=u_type_vec)
